requirements.txt
fastapi
uvicorn
python-dotenv
pydantic
python-multipart
openai>=1.3.0
fastapi
uvicorn[standard]
python-dotenv
python-multipart
pymupdf
openai
groq
openai-whisper
torch
gTTS
soundfile
numpy
starlette
